Comment: 1. 調整 專案的檔案路徑 2. coordinator 整合成一個
