+++ Shpping/Domain/Entity/Model/Endpoint.swift
