+++ Shpping/Domain/Entity/Model/ErrorTracker.swift
