+++ Shpping/Domain/Entity/Model/ProductListEndpoint.swift
