+++ Shpping/Feature/HistoryOrder/HistoryOrderCoordinator.swift
@@ -1,0 +1,12 @@
+//
+//  HistoryOrderCoordinator.swift
+//  Shpping
+//
+//  Created by Jie liang Huang on 2022/6/24.
+//
+
+import RxSwift
+
+protocol HistoryOrderCoordinatorProtocol {
+    func dismiss() -> Observable<Void>
+}
