--- Shpping/View/ProductList/ProductCoordinator.swift
@@ -1,57 +1,0 @@
-//
-//  ProductCoordinator.swift
-//  Shpping
-//
-//  Created by Jie liang Huang on 2022/6/23.
-//
-
-import Foundation
-import RxSwift
-
-protocol ProductCoordinatorProcotol {
-    func showDetailPage(_ model: ShopItemsViewModel) -> Observable<Void>
-    func showHistoryView() -> Observable<Void>
-}
-
-final class ProductCoordinator {
-    weak var viewCotroller: ProductListViewController?
-}
-
-extension ProductCoordinator: ProductCoordinatorProcotol {
-    func showDetailPage(_ model: ShopItemsViewModel) -> Observable<Void> {
-        return .create { [weak self] subscriber in
-            guard let viewController = self?.viewCotroller else {
-                subscriber.onCompleted()
-                return Disposables.create()
-            }
-            let useCase = Domain.ProductDetail(model: model)
-            let coordinator = ProductDetailCoordinator()
-            let viewModel = ProductDetailViewModel(useCase: useCase, coordinator: coordinator)
-            let detailView = ProductDetailViewController(viewModel)
-            coordinator.viewController = detailView
-            viewController.navigationController?.pushViewController(detailView, animated: true)
-            subscriber.onNext(())
-            subscriber.onCompleted()
-            return Disposables.create()
-        }
-    }
-    func showHistoryView() -> Observable<Void> {
-        return .create { [weak self] subscriber in
-            guard let viewController = self?.viewCotroller else {
-                subscriber.onCompleted()
-                return Disposables.create()
-            }
-            let useCase = Domain.HistoryOrder()
-            let coordinator = HistoryOrderCoordinator()
-            let viewModel = HistoryOrderViewModel(useCase: useCase, coordinator: coordinator)
-            let history = HistoryOrderViewController(viewModel: viewModel)
-            history.modalPresentationStyle = .fullScreen
-            coordinator.viewController = history
-            viewController.present(history, animated: true, completion: {
-                subscriber.onNext(())
-                subscriber.onCompleted()
-            })
-            return Disposables.create()
-        }
-    }
-}
