--- Shpping/ViewController.swift
@@ -1,20 +1,0 @@
-//
-//  ViewController.swift
-//  Shpping
-//
-//  Created by Jie liang Huang on 2022/6/22.
-//
-
-import UIKit
-
-class ViewController: UIViewController {
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        // Do any additional setup after loading the view.
-        view.backgroundColor = .white
-    }
-
-
-}
-
